Version 0.0.8: - Make the dark theme the default theme - Repurpose the "Pattern" section for "Page Title" matching (All of the patterns from old config are still valid) - Changed the way the status number are displayed (with documentation) - Update the popup page design to be more compact. - Auto close the popup when the settings page are opened - Code improvement

+++ src/background/background.js
@@ -82,16 +82,17 @@ browser.tabs.onUpdated.addListener((tabId, _0, tab) => {
             });
         }
         if (urlList.some(url => url == new URL(tab.url).hostname) ||
-            tabIdList.some(id => id == tabId) || patternList.some(pattern => {
-            switch (pattern.type) {
-                case "string":
-                    return tab.url.indexOf(`${pattern.pattern}`) != -1;
-                case "regex":
-                    return tab.url.match(pattern.pattern) != null;
-                default:
-                    return false;
-            }
-        })) {
+            tabIdList.some(id => id == tabId) ||
+            patternList.some(pattern => {
+                switch (pattern.type) {
+                    case "string":
+                        return tab.title.indexOf(`${pattern.pattern}`) != -1;
+                    case "regex":
+                        return tab.title.match(pattern.pattern) != null;
+                    default:
+                        return false;
+                }
+            })) {
             await browser.history.deleteUrl({ url: tab.url });
             resolve(true);
         }
@@ -106,26 +107,28 @@ browser.tabs.onActivated.addListener(async (e) => {
     const result = await storage().get("nohistory_urlList");
     let urlList = result.nohistory_urlList || [];
     const cond_2 = urlList.some(url => url == new URL(tabs[0].url).hostname);
-    var color = [0, 0, 0, 0];
-    var rating = "";
-    if (!cond_1 && !cond_2) {
-        color = [255, 0, 0, 255];
-        rating = "00";
-    }
-    else if (cond_1 && !cond_2) {
-        color = [255, 128, 0, 255];
-        rating = "01";
-    }
-    else if (!cond_1 && cond_2) {
-        color = [255, 128, 0, 255];
-        rating = "10";
-    }
-    else {
-        color = [0, 128, 0, 255];
-        rating = "11";
-    }
-    browser.browserAction.setBadgeText({ text: rating });
-    browser.browserAction.setBadgeBackgroundColor({ color: color });
+    var pattern = await storage().get("nohistory_patternList");
+    var patternList = pattern.nohistory_patternList || [];
+    const cond_3 = patternList.some(pattern => {
+        switch (pattern.type) {
+            case "string":
+                return tabs[0].title.indexOf(`${pattern.pattern}`) != -1;
+            case "regex":
+                return tabs[0].title.match(pattern.pattern) != null;
+            default:
+                return false;
+        }
+    });
+    var rating = [0, 0, 0];
+    if (cond_2)
+        rating[0] = 1;
+    if (cond_1)
+        rating[1] = 1;
+    if (cond_3)
+        rating[2] = 1;
+    var final_rating = rating.join("");
+    browser.browserAction.setBadgeText({ text: final_rating });
+    browser.browserAction.setBadgeBackgroundColor({ color: [0, 0, 0, 255] });
     browser.browserAction.setBadgeTextColor({ color: [255, 255, 255, 255] });
 });
 //# sourceMappingURL=background.js.map
